+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Adds support for the `astro:content` and `astro:assets` modules inside Bookshop components.
+* Astro Bookshop will now use your configured Vite plugins when building components.
 
 ## v3.8.2 (December 5, 2023)